requirements.txt
streamlit
pandas
numpy
requests
tensorflow
google-genai